style(utils): separate validate func for clean code
