+++ src/Referee.js
@@ -45,11 +45,12 @@ class Referee {
   printScore () {
     if (this.strike === 3) {
       Console.print('3스트라이크');
-    } else if (this.strike === 0 && this.ball === 0) {
+      Console.print('3개의 숫자를 모두 맞히셨습니다! 게임 종료');
+    } else if (!this.strike && !this.ball) {
       Console.print('낫싱');
-    } else if (this.strike === 0) {
+    } else if (!this.strike) {
       Console.print(`${this.ball}볼`);
-    } else if (this.ball === 0) {
+    } else if (!this.ball) {
       Console.print(`${this.strike}스트라이크`);
     } else {
       Console.print(`${this.ball}볼 ${this.strike}스트라이크`);
